pyproject.toml
[project]
name = "pod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-vlc>=3.0.21203",
    "requests>=2.32.3",
    "textual>=3.1.1",
]
